그래프 옵션 훅으로 분리
